ref: Remove code from bootstrap
